apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.3.7
langchain-google-genai==2.0.5
google-generativeai==0.8.3
